third_party/WebKit/LayoutTests/external/wpt/html/rendering/non-replaced-elements/flow-content-0/dialog-display-expected.txt

This is a testharness.js-based test.
FAIL dialog: display Uncaught Error: assert_equals: expected "block" but got "inline"
Harness: the test ran to completion.
